broke apart left panel components
also updated items as per demo call w/Tomasz
